sparrow-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends.build"

[project]
name = "opensparrow-core"
version = "0.1.0"
description = "OpenSparrow Control Plane — Lightweight AI agent backend for small teams"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
openai = ["openai>=1.50.0"]
anthropic = ["anthropic>=0.34.0"]
telegram = ["python-telegram-bot>=21.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "ruff>=0.6.0"]

[project.scripts]
sparrow-core = "sparrow.main:cli_entry"
